Remove atom from roll-all-18s and flip cond branches
Atoms are ugly and there isn't a need for one since there is already a `loop` construct to hold state for us.

I also flipped the first two branches of the conditional. This should allow the 1 hundred millionth roll to hit all 18s. Before, this roll could have hit all 18s but it quit early.